Core tests 25 Push
Moved applicant and recruiter tests to correct apps.
Left Extra debugging code in the tests in case of regressions.
Set to push
